Renaming of modules to adhere to naming convention.
