+++ src/View/mainpage/ClinicalDataDisplay.py
