+++ src/View/mainpage/ClinicalDataForm.py
