+++ src/View/mainpage/DeleteROIWindow.py
