+++ src/View/mainpage/PatientBar.py
