+++ src/View/mainpage/StructureInformation.py
