+++ src/View/mainpage/resources.qrc
