+++ src/View/mainpage/resources_rc.py
